requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.57.1
anyio==4.9.0
async-timeout==5.0.1
asyncer==0.0.8
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.2
captcha==0.7.1
certifi==2025.6.15
cffi==1.17.1
chainlit==2.6.0
charset-normalizer==3.4.2
chevron==0.14.0
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
exceptiongroup==1.3.0
extra-streamlit-components==0.1.80
fastapi==0.115.14
filelock==3.18.0
filetype==1.2.0
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.175.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
griffe==1.7.3
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.2
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
Lazify==0.4.0
literalai==0.1.201
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mcp==1.10.1
monotonic==1.6
multidict==6.6.3
mypy_extensions==1.1.0
narwhals==1.45.0
nest-asyncio==1.6.0
numpy==2.2.6
openai==1.93.0
openai-agents==0.1.0
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-alephalpha==0.40.14
opentelemetry-instrumentation-anthropic==0.40.14
opentelemetry-instrumentation-bedrock==0.40.14
opentelemetry-instrumentation-chromadb==0.40.14
opentelemetry-instrumentation-cohere==0.40.14
opentelemetry-instrumentation-crewai==0.40.14
opentelemetry-instrumentation-google-generativeai==0.40.14
opentelemetry-instrumentation-groq==0.40.14
opentelemetry-instrumentation-haystack==0.40.14
opentelemetry-instrumentation-lancedb==0.40.14
opentelemetry-instrumentation-langchain==0.40.14
opentelemetry-instrumentation-llamaindex==0.40.14
opentelemetry-instrumentation-logging==0.55b1
opentelemetry-instrumentation-marqo==0.40.14
opentelemetry-instrumentation-mcp==0.40.14
opentelemetry-instrumentation-milvus==0.40.14
opentelemetry-instrumentation-mistralai==0.40.14
opentelemetry-instrumentation-ollama==0.40.14
opentelemetry-instrumentation-openai==0.40.14
opentelemetry-instrumentation-pinecone==0.40.14
opentelemetry-instrumentation-qdrant==0.40.14
opentelemetry-instrumentation-redis==0.55b1
opentelemetry-instrumentation-replicate==0.40.14
opentelemetry-instrumentation-requests==0.55b1
opentelemetry-instrumentation-sagemaker==0.40.14
opentelemetry-instrumentation-sqlalchemy==0.55b1
opentelemetry-instrumentation-threading==0.55b1
opentelemetry-instrumentation-together==0.40.14
opentelemetry-instrumentation-transformers==0.40.14
opentelemetry-instrumentation-urllib3==0.55b1
opentelemetry-instrumentation-vertexai==0.40.14
opentelemetry-instrumentation-watsonx==0.40.14
opentelemetry-instrumentation-weaviate==0.40.14
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-semantic-conventions-ai==0.4.9
opentelemetry-util-http==0.55b1
packaging==25.0
pandas==2.3.0
pillow==11.3.0
posthog==3.25.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-multipart==0.0.18
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.2
requests==2.32.4
rpds-py==0.26.0
rsa==4.9.1
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sse-starlette==2.3.6
starlette==0.46.2
streamlit==1.46.1
streamlit-authenticator==0.4.2
syncer==2.0.3
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.2
toml==0.10.2
tomli==2.2.1
tornado==6.5.1
tqdm==4.67.1
traceloop-sdk==0.40.14
types-requests==2.32.4.20250611
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uptrace==1.34.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
watchfiles==0.20.0
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.1
zipp==3.23.0
